requirements.txt
streamlit
folium
streamlit-folium
pillow
json
timm
torch
pandas
torchvision